Change repository for CTI Git Course
Added the CTI_README.md to provide additional instructions
on launching the website for new users. Changed the main
html page to show 3 placeholder images that can be replaced
by students.

Note that this site uses Bootstrap3 and not Bootstrap5. Working
on changes to update the website to more modern web technologies.